# docs/api/openapi/linksee-v1.yaml
openapi: 3.0.3
info:
  title: Linksee Teaching Collaboration API
  version: 1.0.0
  description: >
    Linksee 教学项目协作与过程评价平台的 P0 API schema。
    本文件与 docs/api/contracts/communication-contract-v1.md 保持一致，
    优先覆盖课程项目、阶段要求、小组讨论、阶段提交、教师反馈和课程看板。
servers:
  - url: /api/v1
    description: Linksee API v1
...
paths:
...
  /courses/{courseId}/announcements:
    post:
      tags: [Groups]
      summary: 课程公告（仅课程管理侧）
      operationId: createCourseAnnouncement
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        "201": { description: 创建成功 }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /courses/{courseId}/messages/search:
    get:
      tags: [Groups]
      summary: 检索课程群消息
      operationId: searchCourseMessages
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { description: 检索成功 }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /courses/{courseId}/messages/{messageId}:
    patch:
      tags: [Groups]
      summary: 编辑课程群消息
      operationId: editCourseMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: 更新成功 }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Groups]
      summary: 删除课程群消息
      operationId: deleteCourseMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: 删除成功 }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
      description: 课程 ID。
...
  responses:
    Unauthenticated:
      description: 未认证或 token 无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: UNAUTHENTICATED
            message: Authentication token is missing or invalid.
            details: null
            requestId: trace_20260430_001
    Forbidden:
      description: 当前用户无权访问该资源或执行该操作
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: FORBIDDEN
            message: Current user is not allowed to perform this operation.
            details: null
            requestId: trace_20260430_002
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: NOT_FOUND
            message: Requested resource was not found.
            details: null
            requestId: trace_20260430_003
...
    ValidationFailed:
      description: 请求字段校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: VALIDATION_FAILED
            message: Request payload validation failed.
            details: null
            requestId: trace_20260430_005
  schemas:
    ErrorResponse:
      type: object
      required:
        - ok
        - code
        - message
      properties:
        ok:
          type: boolean
          enum: [false]
        code:
          type: string
          enum:
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - VALIDATION_FAILED
            - INTERNAL_ERROR
            - FORCE_CHANGE_PASSWORD
        message:
          type: string
        details:
          nullable: true
        requestId:
          type: string
          nullable: true
...